+++ src/effects/breathe.hh
@@ -9,9 +9,9 @@
 
 class Breathe : public Effect {
     public:
-        explicit Breathe(Region &region, uint8_t steps = 60, uint16_t timeIn = 250, uint16_t timeOut = 250, uint16_t timeOff = 100);
+        explicit Breathe(Region &region, uint8_t steps = 60, uint32_t timeIn = 250, uint32_t timeOut = 250, uint32_t timeOff = 100);
         void init() override;
-        uint64_t loop() override;
+        uint32_t loop() override;
     private:
         uint8_t steps;
         uint16_t timeIn;
