+++ src/effects/sequential_trails.hh
@@ -9,17 +9,17 @@
 
 class SequentialTrails : public Effect {
 public:
-    explicit SequentialTrails(Region &region, Color color, uint8_t length, uint8_t count, uint16_t loopTime, bool circular = true);
+    explicit SequentialTrails(Region &region, Color color, uint8_t length, uint8_t count, uint32_t loopTime, bool circular = true);
 
     void init() override;
-    uint64_t loop() override;
+    uint32_t loop() override;
 
     ~SequentialTrails();
 private:
     const Color color;
     const uint8_t length;
     const uint8_t count;
-    const uint16_t loopTime;
+    const uint32_t loopTime;
     const bool circular;
     uint8_t headDist;
     uint8_t *heads = (uint8_t *)malloc((count + 1) * sizeof(uint8_t));
