+++ src/effects/spectrum_cycle.hh
@@ -10,14 +10,14 @@
 
 class SpectrumCycle : public Effect {
 public:
-    explicit SpectrumCycle(Region &region, Color base, uint16_t loopTime = 1000, uint16_t fadeTime = 0);
+    explicit SpectrumCycle(Region &region, Color base, uint32_t loopTime = 1000, uint32_t fadeTime = 0);
     void init() override;
-    uint64_t loop() override;
+    uint32_t loop() override;
 private:
     const Color base;
     ColorCycling active;
-    const uint16_t loopTime;
-    const uint16_t fadeTime;
+    const uint32_t loopTime;
+    const uint32_t fadeTime;
     double curFade;
     bool fadingIn;
 };
