+++ src/effects/static_color.cc
@@ -4,7 +4,7 @@
 
 #include "static_color.hh"
 
-StaticColor::StaticColor(Region &region, Color color, uint16_t delay) : Effect(region), color(color), delay(delay) {
+StaticColor::StaticColor(Region &region, Color color, uint32_t delay) : Effect(region), color(color), delay(delay) {
 
 }
 
@@ -12,7 +12,7 @@ void StaticColor::init() {
 
 }
 
-uint64_t StaticColor::loop() {
+uint32_t StaticColor::loop() {
     for (uint16_t i = 0; i < region.size; i++) {
         region.set(i, color);
     }
